+++ app/controllers/account_controller.rb
@@ -32,6 +32,6 @@ class AccountController < ApplicationController
   private
 
   def user_params
-    params.require(:user).permit(:first_name, :last_name, :email, :emails_enabled)
+    params.require(:user).permit(:first_name, :last_name, :email, :emails_enabled, :calendar_view)
   end
 end