+++ db/migrate/20160801151245_add_calendar_view_to_users.rb
@@ -1,0 +1,5 @@
+class AddCalendarViewToUsers < ActiveRecord::Migration[5.0]
+  def change
+    add_column :users, :calendar_view, :string, null: false, default: 'month'
+  end
+end